+++ .gitignore
@@ -1,5 +1,3 @@
-lib/dependencygraph.json
-.rpt2_cache
 node_modules
 dist
 *css.d.ts
