+++ package.json
@@ -10,19 +10,15 @@
   ],
   "license": "Apache-2.0",
   "scripts": {
-    "build:markup": "node ./lib/render_static.js",
     "gzip": "gzip -k9nr dist",
-    "build": "rollup -c && npm run build:markup && npm run gzip",
-    "watch:rollup": "rollup -c -w",
-    "watch": "run-p build:markup watch:rollup",
+    "build": "rollup -c && npm run gzip",
+    "watch": "rollup -c -w",
     "serve": "http-server dist -c-1 -g",
     "fmt": "prettier --write 'src/**/*.{ts,tsx,css}' *.{js,json,md} 'lib/**/*.js' && npm run lintfix",
     "lint": "tslint -t verbose -c tslint.json 'src/**/*.{ts,tsx}'",
     "lintfix": "tslint -t verbose -c tslint.json --fix 'src/**/*.{ts,tsx}'",
     "fmt_test": "test $(prettier -l 'src/**/*.{ts,tsx,css}' *.{js,json,md} 'lib/**/*.js' | wc -l) -eq 0",
-    "test": "npm run fmt_test && npm run lint && npm run build",
-    "tmpbuild": "rollup -c",
-    "tmpbuilddebug": "node --inspect-brk node_modules/.bin/rollup -c"
+    "test": "npm run fmt_test && npm run lint && npm run build"
   },
   "husky": {
     "hooks": {
