+++ src/main/services/preact-canvas/components/game/index.tsx
@@ -14,12 +14,12 @@ import { Remote } from "comlink/src/comlink";
 import { Component, h } from "preact";
 import { Animator } from "src/main/rendering/animator";
 import { Renderer } from "src/main/rendering/renderer";
-import StateService from "../../../../../worker/state-service";
 import { submitTime } from "src/main/services/state/best-times";
 import { supportsVibration } from "src/main/services/state/vibration-preference";
 import { bind } from "src/utils/bind";
 import { StateChange } from "../../../../../worker/gamelogic";
 import { Cell, PlayMode } from "../../../../../worker/gamelogic/types";
+import StateService from "../../../../../worker/state-service";
 import { vibrationLength } from "../../../../utils/constants";
 import initFocusHandling from "../../../../utils/focus-visible";
 import { isFeaturePhone } from "../../../../utils/static-display";
